refactor: modular DetailView components
